chore: Add opinion-service configuration files and classes

+++ opinion-service/src/main/java/edu/bethlehem/opinion_service/Opinion/OpinionService.java
@@ -1,4 +1,4 @@
-package edu.bethlehem.post_service.Opinion;
+package edu.bethlehem.opinion_service.Opinion;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import edu.bethlehem.post_service.Post.PostNotFoundException;
 import edu.bethlehem.post_service.Post.PostRepository;
+import edu.bethlehem.post_service.Security.JwtServiceProxy;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
@@ -19,7 +20,7 @@ public class OpinionService {
 
         private final PostRepository postRepository;
         private final OpinionRepository opinionRepository;
-
+        private final JwtServiceProxy jwtServiceProxy;
         private final OpinionModelAssembler assembler;
 
         public Opinion convertOpinionDtoToOpinionEntity(OpinionDTO opinionDTO, Long userId) {
@@ -45,14 +46,20 @@ public class OpinionService {
                                 .collect(Collectors.toList());
         }
 
-        public EntityModel<Opinion> postOpinion(OpinionDTO newOpinionDTO, Long userId) {
+        public EntityModel<Opinion> postOpinion(OpinionDTO newOpinionDTO, String jwt) {
+                Long userId = jwtServiceProxy.extractUserId(jwt);
                 Opinion newOpinion = convertOpinionDtoToOpinionEntity(newOpinionDTO, userId);
                 return assembler.toModel(opinionRepository.save(newOpinion));
         }
 
-        public EntityModel<Opinion> updateOpinionPartially(Long opinionId, OpinionPatchDTO opinionPatchDTO) {
+        public EntityModel<Opinion> updateOpinionPartially(Long opinionId, OpinionPatchDTO opinionPatchDTO,
+                        String jwt) {
+
                 Opinion opinion = opinionRepository.findById(opinionId)
                                 .orElseThrow(() -> new OpinionNotFoundException(opinionId, HttpStatus.NOT_FOUND));
+
+                if (jwtServiceProxy.extractUserId(jwt) != opinion.getUserId())
+                        throw new OpinionNotFoundException(opinionId, HttpStatus.FORBIDDEN);
 
                 if (opinionPatchDTO.getContent() != null)
                         opinion.setContent(opinionPatchDTO.getContent());
@@ -61,7 +68,10 @@ public class OpinionService {
 
         }
 
-        public void deleteOpinion(Long id) {
+        public void deleteOpinion(Long id, String jwt) {
+
+                if (jwtServiceProxy.extractUserId(jwt) != opinionRepository.findById(id).get().getUserId())
+                        throw new OpinionNotFoundException(id, HttpStatus.FORBIDDEN);
 
                 opinionRepository.deleteById(id);
